refactor: change folder structure

+++ Server/src/main/java/com/example/cardapio/dto/foodDto/FoodReponseDTO.java
@@ -2,7 +2,7 @@ package com.example.cardapio.dto.foodDto;
 
 import com.example.cardapio.models.Food;
 
-public record FoodReponseDTO(long id, String title, String image, Integer price ) {
+public record FoodReponseDTO(long id, String title, String image, Double price ) {
   public FoodReponseDTO(Food food){
     this(food.getId(), food.getTitle(), food.getImage(), food.getPrice());
   }
